Make class for API instantiable
